feat: 테스트를 위한 MockLinkedList 타입 생성 #1
- 접근제어 추가
