+++ BankManagerConsoleApp/BankManagerConsoleAppTests/LinkedListTests.swift
@@ -1,59 +1,10 @@
 import XCTest
 
-final class LinkedList<Element> {
-    final class Node<Element> {
-        var value: Element
-        var next: Node<Element>?
-        
-        init(value: Element) {
-            self.value = value
-        }
-    }
-    
-    var head: Node<Element>?
-    var tail: Node<Element>?
-    
-    var first: Element? {
-        return head?.value
-    }
-    
-    var isEmpty: Bool {
-        return head == nil
-    }
-    
-    func append(_ value: Element) {
-        let newNode = Node(value: value)
-
-        if isEmpty {
-            head = newNode
-            tail = newNode
-        } else {
-            tail?.next = newNode
-            tail = newNode
-        }
-    }
-    
-    func removeFirst() -> Element? {
-        guard isEmpty == false else {
-            return nil
-        }
-        let result = head?.value
-        head = head?.next
-        
-        return result
-    }
-    
-    func removeAll() {
-        head = nil
-        tail = nil
-    }
-}
-
 class LinkedListTests: XCTestCase {
-    var sut: LinkedList<Int>!
+    var sut: MockLinkedList<Int>!
     
     override func setUpWithError() throws {
-        sut = LinkedList()
+        sut = MockLinkedList()
     }
     
     override func tearDownWithError() throws {
